// types/transactions.ts
import { z } from "zod"

export const transactionSchema = z.object({
  id: z.number(),
  date: z.string(),
  orNo: z.string(),
  acNum: z.string(),
  name: z.string(),
  amount: z.number(),
  bankCode: z.string(),
  country: z.string(),
  isCash: z.boolean(),
  checkNo: z.string(),
  checkDate: z.string(),
  userId: z.string(),
  stat: z.string(),
  type: z.string(),
  email: z.string(),
  employmentStatus: z.string(),
  isFlagged: z.boolean(),
  flagReason: z.string(),
  balance: z.number(),
  contactChanges: z.string(),
  mot: z.number(),
  purpose: z.string(),
  productType: z.string(),
  idType: z.string(),
  idNo: z.string(),
  sourceFund: z.string(),
  currencyCode: z.string(),
  cityCode: z.string(),
})